// src/app/atendimento-paciente/ficha-consulta-retorno/shared/model/avaliacao-massa-muscular-corporea.model.ts
export class AvaliacaoMassaMuscularCorporea {

    id: number;
    circunferenciaCinturaCentimetros: string;
	circunferenciaBracoCentimentros: string;
	circunferenciaCoxaCentimetros: string;
	circunferenciaPanturrilhaCentimetros: string;
	circunferenciaPunhoCentrimetros: string;
	massaMuscularKg: string;
	massaMuscularPorcentagem: string;
	indiceMassaMuscularKgMetroQuadrado: string;
}